+++ lib/legion/extensions/ssh.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 
 require 'legion/extensions/ssh/version'
+require 'legion/extensions/ssh/helpers/connection'
+require 'legion/extensions/ssh/runners/command'
+require 'legion/extensions/ssh/runners/state'
+require 'legion/extensions/ssh/client'
 
 module Legion
   module Extensions