+++ lib/legion/extensions/ssh/runners/command.rb
@@ -24,7 +24,10 @@ module Legion
             final_results
           end
 
-          include Legion::Extensions::Helpers::Lex
+          alias run_multiple run_mulitple
+
+          include Legion::Extensions::Helpers::Lex if Legion::Extensions.const_defined?(:Helpers) &&
+                                                      Legion::Extensions::Helpers.const_defined?(:Lex)
         end
       end
     end